issue #41
First pass at factoring out the code that fixes the up failed service
request-id.

The Chicago specific code has been moved to fix_request_id_chi, which
is assigned to fix_request_id, which is the function that is actually
called in show_request().  Any other cities need to write an
equivalent to fix_request_id_chi and assign it to fix_request_id
(presumably, there is a way to make this more pluggable/configurable).